att ladda upp via webbsidan är en dum idé för nu behövde jag fixa detta
