--- config.py
@@ -1,8 +1,0 @@
-import os
-basedir = os.path.abspath(os.path.dirname(__file__))
-
-class Config(object):
-    SQLALCHEMY_DATABASE_URI = os.environ.get or \
-        "sqlite:///" + os.path.join(basedir, "app.db")
-    
-    SQLALCHEMY_TRACK_MODIFICATIONS = False
