+++ frack/app/db_interface.py
@@ -1,11 +1,14 @@
 import pymysql
 
 
+try:
+    conn = pymysql.connect(host="127.0.0.1", port=3306, user="root", passwd="", db="mysql", autocommit= True) 
+    conn.autocommit(True)
+    cursor = conn.cursor()
+except:
+    print("kunde inte nå databasen!")
 
-conn = pymysql.connect(host="127.0.0.1", port=3306, user="root", passwd="", db="mysql", autocommit= True) 
-conn.autocommit(True)
-cursor = conn.cursor()
-
+#hämta nyheter; antigen specifikt id eller alla
 def get_news(id):
     if id is not None:
 
@@ -16,7 +19,7 @@ def get_news(id):
             
         return j_dict
     else:
-        sql = "SELECT * FROM mtgn18.news WHERE 'id' IS NOT NULL"
+        sql = "SELECT * FROM mtgn18.news WHERE 'id' IS NOT NULL ORDER BY id DESC"
         cursor.execute(sql, id)
         result = cursor.fetchall()
 
@@ -28,12 +31,18 @@ def get_news(id):
             
         return news_list
 
+#spara ny nyhet i DB
 def add_news(dict):
         sql = "INSERT INTO mtgn18.news (author, headline, text, tags) VALUES(%s, %s, %s, %s)"
         cursor.execute(sql, (dict["author"], dict["headline"], dict["text"], dict["tags"]))
         
 
-
+#radera entry från DB (försvinner för alltid!)
 def delete_news(id):
     sql = "DELETE from mtgn18.news WHERE id = %s"
     cursor.execute(sql, id)
+
+#uppdatera entry i databas (edit)
+def edit_news(id, dict):
+    sql = "UPDATE mtgn18.news SET (author, headline, text, tags) = (%s, %s, %s, %s) WHERE id = %s "
+    cursor.execute(sql, (dict["author"], dict["headline"], dict["text"], dict["tags"], id))
