+++ frack/app/webapp.py
@@ -1,8 +1,10 @@
 from flask import Flask
 import os
 
-app = Flask(__name__)
+UPLOAD_FOLDER = os.path.join(os.getcwd(), "static", "Schmeck", "media")
 
+app = Flask(__name__)
+app.config["UPLOAD_FOLDER"] = UPLOAD_FOLDER
 
 
 from app import routes
