+++ frack/frack.py
@@ -1,4 +1,4 @@
-from app.webapp import app
+from app import app
 
 if __name__ == "__main__":
     app.run(debug = True)
