+++ frack/requirements.txt
@@ -1,5 +1,6 @@
 click==6.7
 Flask==1.0.2
+Flask-SQLAlchemy==2.3.2
 itsdangerous==0.24
 Jinja2==2.10
 MarkupSafe==1.0
