--- requirements.txt
@@ -1,8 +1,0 @@
-click==6.7
-Flask==1.0.2
-Flask-SQLAlchemy==2.3.2
-itsdangerous==0.24
-Jinja2==2.10
-MarkupSafe==1.0
-PyMySQL==0.8.1
-Werkzeug==0.14.1
